requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
